Sets all perl modules to versions to 0.9.0
I'm not sure how I want to handle these version strings, as
we currently are not present on CPAN. So this value seems not
important and seems to have no real use so far. Not sure how
I want to increase these version numbers (rethink this when
we really have to implement a proper release process).