Fixes css uri handling for htc files (IE behavior)
From: http://css3pie.com/documentation/known-issues/#relative-paths

IE interprets the URL for the behavior property relative to
the source HTML document, rather than relative to the CSS file
like every other CSS property. This makes invoking the PIE
behavior inconvenient, because the URL has to either be:

Absolute from the domain root — this makes the CSS not easily
moveable between directories — or,
Relative to the HTML document — this makes the CSS not easily
reusable between different HTML files.

We only support the first method. You can include these files
as all other urls in the css, but it will always be exported
absolute to the webroot.